[Refact] #66 - 모델&dto refactoring
- 기존에 model 내부에 있던 req&res를 dto로 refact
- model과 req&res 분리
