+++ app/src/main/java/org/cazait/data/api/AuthService.kt
@@ -1,8 +1,8 @@
 package org.cazait.data.api
 
-import org.cazait.data.model.request.SignInReq
+import org.cazait.data.dto.request.SignInReq
 import org.cazait.data.dto.response.RefreshTokenRes
-import org.cazait.data.model.response.SignInRes
+import org.cazait.data.dto.response.SignInRes
 import retrofit2.Call
 import retrofit2.http.*
