+++ app/src/main/java/org/cazait/data/api/CafeService.kt
@@ -2,8 +2,8 @@ package org.cazait.data.api
 
 import org.cazait.data.dto.response.ListCafesRes
 import org.cazait.data.dto.response.ListFavoritesRes
-import org.cazait.data.model.response.CafeMenuRes
-import org.cazait.data.model.response.CafeReviewRes
+import org.cazait.data.dto.response.CafeMenuRes
+import org.cazait.data.dto.response.CafeReviewRes
 import retrofit2.Call
 import retrofit2.http.GET
 import retrofit2.http.Path
