+++ app/src/main/java/org/cazait/data/dto/request/SignInReq.kt
@@ -1,4 +1,4 @@
-package org.cazait.data.model.request
+package org.cazait.data.dto.request
 
 data class SignInReq(
     val email: String,
