+++ app/src/main/java/org/cazait/data/dto/response/IsNicknameDupRes.kt
@@ -1,4 +1,4 @@
-package org.cazait.data.model.response
+package org.cazait.data.dto.response
 
 data class IsNicknameDupRes(
     val code: Int,
