+++ app/src/main/java/org/cazait/data/remote/cafe/CafeInfoRemoteDataSource.kt
@@ -1,8 +1,8 @@
 package org.cazait.data.remote.cafe
 
 import org.cazait.data.Resource
-import org.cazait.data.model.response.CafeMenuRes
-import org.cazait.data.model.response.CafeReviewRes
+import org.cazait.data.dto.response.CafeMenuRes
+import org.cazait.data.dto.response.CafeReviewRes
 
 interface CafeInfoRemoteDataSource {
     suspend fun getMenus(cafeId: Long): Resource<CafeMenuRes>
