+++ app/src/main/java/org/cazait/data/repository/auth/AuthRepository.kt
@@ -2,8 +2,8 @@ package org.cazait.data.repository.auth
 
 import kotlinx.coroutines.flow.Flow
 import org.cazait.data.Resource
-import org.cazait.data.model.request.SignInReq
-import org.cazait.data.model.response.SignInRes
+import org.cazait.data.dto.request.SignInReq
+import org.cazait.data.dto.response.SignInRes
 
 interface AuthRepository {
     suspend fun refreshToken()
