+++ app/src/main/java/org/cazait/data/repository/auth/AuthRepositoryImpl.kt
@@ -4,8 +4,8 @@ import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.flow
 import kotlinx.coroutines.flow.flowOn
 import org.cazait.data.Resource
-import org.cazait.data.model.request.SignInReq
-import org.cazait.data.model.response.SignInRes
+import org.cazait.data.dto.request.SignInReq
+import org.cazait.data.dto.response.SignInRes
 import org.cazait.data.remote.auth.AuthRemoteData
 import javax.inject.Inject
 import kotlin.coroutines.CoroutineContext
