+++ app/src/main/java/org/cazait/data/repository/cafe/CafeRepositoryImpl.kt
@@ -7,8 +7,8 @@ import org.cazait.data.Resource
 import org.cazait.data.dto.request.ListCafesReq
 import org.cazait.data.dto.response.ListCafesRes
 import org.cazait.data.dto.response.ListFavoritesRes
-import org.cazait.data.model.response.CafeMenuRes
-import org.cazait.data.model.response.CafeReviewRes
+import org.cazait.data.dto.response.CafeMenuRes
+import org.cazait.data.dto.response.CafeReviewRes
 import org.cazait.data.remote.cafe.CafeInfoRemoteData
 import org.cazait.data.remote.cafe.CafeListRemoteData
 import javax.inject.Inject
