+++ app/src/main/java/org/cazait/data/repository/users/UserRepository.kt
@@ -2,12 +2,12 @@ package org.cazait.data.repository.users
 
 import kotlinx.coroutines.flow.Flow
 import org.cazait.data.Resource
-import org.cazait.data.model.request.IsEmailDupReq
-import org.cazait.data.model.request.IsNicknameDupReq
-import org.cazait.data.model.request.SignUpReq
-import org.cazait.data.model.response.IsEmailDupRes
-import org.cazait.data.model.response.IsNicknameDupRes
-import org.cazait.data.model.response.SignUpRes
+import org.cazait.data.dto.request.IsEmailDupReq
+import org.cazait.data.dto.request.IsNicknameDupReq
+import org.cazait.data.dto.request.SignUpReq
+import org.cazait.data.dto.response.IsEmailDupRes
+import org.cazait.data.dto.response.IsNicknameDupRes
+import org.cazait.data.dto.response.SignUpRes
 
 interface UserRepository {
     suspend fun signUp(body: SignUpReq): Flow<Resource<SignUpRes>>
