+++ app/src/main/java/org/cazait/data/repository/users/UserRepositoryImpl.kt
@@ -4,12 +4,12 @@ import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.flow
 import kotlinx.coroutines.flow.flowOn
 import org.cazait.data.Resource
-import org.cazait.data.model.request.IsEmailDupReq
-import org.cazait.data.model.request.IsNicknameDupReq
-import org.cazait.data.model.request.SignUpReq
-import org.cazait.data.model.response.IsEmailDupRes
-import org.cazait.data.model.response.IsNicknameDupRes
-import org.cazait.data.model.response.SignUpRes
+import org.cazait.data.dto.request.IsEmailDupReq
+import org.cazait.data.dto.request.IsNicknameDupReq
+import org.cazait.data.dto.request.SignUpReq
+import org.cazait.data.dto.response.IsEmailDupRes
+import org.cazait.data.dto.response.IsNicknameDupRes
+import org.cazait.data.dto.response.SignUpRes
 import org.cazait.data.remote.user.UserRemoteData
 import javax.inject.Inject
 import kotlin.coroutines.CoroutineContext
