+++ app/src/main/java/org/cazait/ui/component/cafeinfo/CafeInfoViewModel.kt
@@ -3,9 +3,6 @@ package org.cazait.ui.component.cafeinfo
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import dagger.hilt.android.lifecycle.HiltViewModel
-import org.cazait.data.model.CafeImage
-import org.cazait.data.dto.response.CafeOfCafeList
-import org.cazait.data.model.Cafe
 import org.cazait.ui.base.BaseViewModel
 import javax.inject.Inject
