+++ app/src/main/java/org/cazait/ui/component/cafeinfo/menu/CafeInfoMenuViewModel.kt
@@ -6,7 +6,7 @@ import androidx.lifecycle.viewModelScope
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.launch
 import org.cazait.data.Resource
-import org.cazait.data.model.response.CafeMenuRes
+import org.cazait.data.dto.response.CafeMenuRes
 import org.cazait.data.repository.cafe.CafeRepository
 import org.cazait.ui.base.BaseViewModel
 import javax.inject.Inject
