+++ app/src/main/java/org/cazait/ui/component/signin/SignInViewModel.kt
@@ -8,8 +8,8 @@ import androidx.lifecycle.viewModelScope
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.launch
 import org.cazait.data.Resource
-import org.cazait.data.model.request.SignInReq
-import org.cazait.data.model.response.SignInRes
+import org.cazait.data.dto.request.SignInReq
+import org.cazait.data.dto.response.SignInRes
 import org.cazait.data.repository.auth.AuthRepository
 import org.cazait.ui.base.BaseViewModel
 import org.cazait.utils.SingleEvent
