+++ app/src/main/java/org/cazait/ui/component/signup/SignUpViewModel.kt
@@ -6,12 +6,12 @@ import androidx.lifecycle.viewModelScope
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.launch
 import org.cazait.data.Resource
-import org.cazait.data.model.request.IsEmailDupReq
-import org.cazait.data.model.request.IsNicknameDupReq
-import org.cazait.data.model.request.SignUpReq
-import org.cazait.data.model.response.IsEmailDupRes
-import org.cazait.data.model.response.IsNicknameDupRes
-import org.cazait.data.model.response.SignUpRes
+import org.cazait.data.dto.request.IsEmailDupReq
+import org.cazait.data.dto.request.IsNicknameDupReq
+import org.cazait.data.dto.request.SignUpReq
+import org.cazait.data.dto.response.IsEmailDupRes
+import org.cazait.data.dto.response.IsNicknameDupRes
+import org.cazait.data.dto.response.SignUpRes
 import org.cazait.data.repository.users.UserRepository
 import org.cazait.ui.base.BaseViewModel
 import org.cazait.utils.SingleEvent
